backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6
fastapi==0.104.1
h11==0.14.0
idna==3.6
neo4j==5.15.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
pytz==2023.3.post1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
